Fixed api.uploadBooks() and /uploadBooks slightly.
Fixed api.resolveIsbn() and IsbnBook model
to coalesce into an optional (nullable) image field.

Instead of writing `this.method.bind(this)`, in the constructor,
I changed them to `private readonly method = (): void {};` in the class.
Ideally, I'd do this for every method,
but WebStorm doesn't see updates on this in those methods.

Also added more object destructuring to simplify things.