Trying to fix random log outs in Apache.
Narrowed it down to itsdangerous.Signer.derive_key(),
which usually returns the same key for each request from the same user,
but whenever the user is inexplicably logged out,
it returns a different key.

Since the key is different, the calculated signatures are different,
which causes the session to be invalidated and a new session to be generated.
The _paranoid_token still in the session
is actually placed there after the session is invalidated.