Trying to fix random log outs in Apache.
The wsgi script is run twice for some reason,
something I've noticed usually happens in debug mode.

The two different secret_keys are being taken from the two app objects.
Even though hash(app) is the same for both,
they are run in different processes, not just different threads,
so the hashes could be the same because they are different Python processes.

I'm checking to see if debug mode is accidentally set on Apache,
although I explicitly set it to false in the wsgi script.

The other more likely possibility is that Apache
is running multiple processes of the same wsgi script,
meaning there are 2+ entirely different app objects
and that anything accessed in RAM
is not guaranteed to be accessible by any other request.
Only things written to disk or to shared memory, etc. (which I'm not using)
will be guaranteed to be accessible to all requests.
Without synchronization, even those interprocess memories
might not be immediately synchronized if requests in different processes
run in parallel and don't have time to see the other process' changes.

I'm not sure how I can fix this.
Maybe the secret_key set to os.urandom(32) can be read from a file
that is checked to be only written to once.
i.e., read the secret_key from a file, and if it's not there yet,
generate the secret_key and write to it.